pubmed-rag/requirements.txt
flask
biopython
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
google-generativeai
